feat: 作成画面のUX改善
